Update Results.html changed "correct" to "best" and clarified on-time bonus
To mitigate the confusion participants say they experience about "correct" answers, I used "best" instead (as we do in our current manuscript draft). 

Also, participants mention confusion about the on-time bonus (since we have yet to actually schedule sessions). So I added "(if this session was scheduled)" to help our participants realize that that doesn't apply. 

Similarly, I added something like "(if there was a bonus)" to the performance bonus line for when we run discussion conditions WITHOUT bonuses (to mitigate the confusion that those discussion participants experience about the bonus line on this page).